Move `createModelWithLargeSize` to models creation test utils and use it for the new added tests
