+++ src/plugins/intel_npu/tests/functional/common/functions.h
@@ -9,6 +9,8 @@
 // create dummy network for tests
 std::shared_ptr<ov::Model> buildSingleLayerSoftMaxNetwork();
 
+std::shared_ptr<ov::Model> createModelWithLargeSize();
+
 // class encapsulated Platform getting from environmental variable
 class PlatformEnvironment {
 public:
